refactor: updated layout usage to layoutV2 and deprecated the existing layout
