+++ packages/ui/src/components/file-control-bars/file-viewer-control-bar.tsx
@@ -43,7 +43,7 @@ export const FileViewerControlBar: FC<FileViewerControlBarProps> = ({
 
   const RightDetails = () => {
     return (
-      <Layout.Horizontal gap="gap-2" className="items-center">
+      <Layout.Horizontal gap="xs" align="center">
         <span className="text-sm text-cn-foreground-2">{`${fileContent?.split('\n').length || 0} lines`}</span>
         <span className="h-3 border-l border-cn-borders-2" />
         <span className="mr-5 text-sm text-cn-foreground-2">{fileBytesSize}</span>
