+++ packages/ui/src/components/index.ts
@@ -82,7 +82,6 @@ export * from './sorts'
 export * from './entity-form-layout'
 export * as MultiSelectV2 from './multi-select-v2'
 export * from './card-select'
-export * from './layout-v2'
 
 export * as NodeGroup from './node-group'
 export * as ShaBadge from './sha-badge'
