+++ packages/ui/src/views/repo/pull-request/details/components/changes/pull-request-changes.tsx
@@ -338,7 +338,7 @@ const PullRequestAccordion: React.FC<{
             <Accordion.Content className="pb-0">
               <div className="border-t bg-transparent">
                 {(fileDeleted || isDiffTooLarge || fileUnchanged || header?.isBinary) && !showHiddenDiff ? (
-                  <Layout.Vertical gap="space-y-0" className="items-center py-5">
+                  <Layout.Vertical align="center" className="py-5">
                     <Button
                       className="text-cn-foreground-accent"
                       variant="link"
